fixing broken tests after renaming of the files in module
