+++ grid2op/tests/test_attached_envs_compat.py
@@ -13,10 +13,8 @@ import grid2op
 import numpy as np
 
 from grid2op.Space import GridObjects
-from grid2op.Action.powerlineSetAction import PowerlineSetAction
-from grid2op.Action.PlayableAction import PlayableAction
-from grid2op.Observation.completeObservation import CompleteObservation
-from grid2op.Action.dontAct import DontAct
+from grid2op.Action import PowerlineSetAction, DontAct, PlayableAction
+from grid2op.Observation import CompleteObservation
 
 import pdb
